update encrypt callers
